bionic: Prefer /sbin/sh if it exists
Prefer /sbin/sh because in recovery (or generally, not booted to the
main Android system), the system shell and its associated rc file
should not be used.  One reason is that TMPDIR is set to a path that
will likely not be mounted (/data/local/tmp).

Change-Id: I3ec22aa2537a7795baae6e782134d3b05f4b3ace
(cherry picked from commit 9028e6d5c297b02a322bf9f26256dcc64b062e05)